LayoutTests/platform/mac-bigsur-wk2/http/tests/websocket/tests/hybi/deflate-frame-set-bfinal-expected.txt

Test receiving compressed frames with BFINAL = 1.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Sending message: "Hello"
PASS event.data is 'Hello'
Sending message: "World"
PASS event.data is 'World'
Sending message: "Goodbye"
PASS event.data is 'Goodbye'
onclose() was called.
FAIL closeEvent.wasClean should be true. Was false.
PASS successfullyParsed is true

TEST COMPLETE
